+++ packages/flow-core/src/components/f-date-time-picker/f-date-time-picker.test.ts
@@ -1,4 +1,4 @@
-import IconPack from "@cldcvr/flow-system-icon/dist/types/icon-pack";
+import IconPack from "@ollion/flow-system-icon/dist/types/icon-pack";
 import { expect, fixture, html, oneEvent } from "@open-wc/testing";
 
 // importing flow-core components