+++ packages/flow-core/src/components/f-icon-button/f-icon-button.test.ts
@@ -1,7 +1,7 @@
 import { html, fixture, expect } from "@open-wc/testing";
 
 // IconPack to test
-import IconPack from "@cldcvr/flow-system-icon/dist/types/icon-pack";
+import IconPack from "@ollion/flow-system-icon/dist/types/icon-pack";
 // import flow-core elements
 import "@ollion/flow-core";
 import { FIconButton, ConfigUtil } from "@ollion/flow-core";