+++ packages/flow-core/src/components/f-search/f-search.test.ts
@@ -1,5 +1,5 @@
 import { expect, fixture, oneEvent } from "@open-wc/testing";
-import IconPack from "@cldcvr/flow-system-icon/dist/types/icon-pack";
+import IconPack from "@ollion/flow-system-icon/dist/types/icon-pack";
 
 // import flow-core elements
 import "@ollion/flow-core";